Buildable

+++ arch/m68k/sun3x/time.c
@@ -15,6 +15,7 @@
 
 #include <asm/irq.h>
 #include <asm/io.h>
+#include <asm/system.h>
 #include <asm/traps.h>
 #include <asm/sun3x.h>
 #include <asm/sun3ints.h>
@@ -71,7 +72,7 @@ int sun3x_hwclk(int set, struct rtc_time *t)
 	return 0;
 }
 /* Not much we can do here */
-u32 sun3x_gettimeoffset(void)
+unsigned long sun3x_gettimeoffset (void)
 {
     return 0L;
 }
