+++ drivers/ide/buddha.c
@@ -23,7 +23,6 @@
 #include <linux/zorro.h>
 #include <linux/ide.h>
 #include <linux/init.h>
-#include <linux/module.h>
 
 #include <asm/amigahw.h>
 #include <asm/amigaints.h>
@@ -198,7 +197,7 @@ fail_base2:
 				continue;
 			}
 		}	  
-		buddha_board = (unsigned long)ZTWO_VADDR(board);
+		buddha_board = ZTWO_VADDR(board);
 		
 		/* write to BUDDHA_IRQ_MR to enable the board IRQ */
 		/* X-Surf doesn't have this.  IRQs are always on */
